requirements.txt
django>=5.0,<6
djangorestframework>=3.15,<4
requests>=2.32
python-dotenv>=1.0
django-crontab>=0.7
gunicorn>=21.2